refactor: move CONTROL_ACTIONS into turn.ts next to tickTurn
The pause/play/back/forward turn arithmetic was defined inline in the
control route. It is turn-engine domain logic, so it now lives in
src/server/turn.ts alongside tickTurn and can be unit-tested directly.
The control route imports it.
