+++ src/server/turn.ts
@@ -1,4 +1,10 @@
-import { ApiResponse, Game, SetupInformation } from "@fc/types/types";
+import {
+    ApiResponse,
+    ControlAction,
+    ControlAPI,
+    Game,
+    SetupInformation,
+} from "@fc/types/types";
 import { Either, isLeft } from "fp-ts/Either";
 import { MakeLeft, MakeRight } from "@fc/lib/io-ts-helpers";
 
@@ -237,6 +243,147 @@ export function getCurrentPhase(
 
     return MakeRight(possible);
 }
+
+/**
+ * The pause/play/back/forward turn arithmetic for the control desk. Lives next
+ * to tickTurn so it is unit-tested with the rest of the turn maths rather than
+ * only through the HTTP route.
+ */
+export const CONTROL_ACTIONS: Record<ControlAPI["action"], ControlAction> = {
+    pause: (game) => {
+        return MakeRight({
+            active: false,
+            frozenTurn: toApiResponse(
+                { ...game, active: false, frozenTurn: toApiResponse(game) },
+                true,
+            ),
+        });
+    },
+    play: (game) => {
+        if (game.active) {
+            return MakeRight(game);
+        }
+
+        const phaseEnd = new Date();
+        phaseEnd.setSeconds(phaseEnd.getSeconds() + game.frozenTurn.phaseEnd);
+
+        const turnInformation: Game["turnInformation"] = {
+            turnNumber: game.frozenTurn.turnNumber,
+            currentPhase: game.frozenTurn.phase,
+            phaseEnd: phaseEnd.toString(),
+        };
+
+        return MakeRight({
+            active: true,
+            turnInformation,
+        });
+    },
+    "back-phase": (game) => {
+        const turnInformation = game.turnInformation;
+
+        const currentPhase = turnInformation.currentPhase;
+
+        let newPhase = currentPhase - 1;
+        let turnNumber = turnInformation.turnNumber;
+
+        if (currentPhase == 1) {
+            if (turnNumber == 1) {
+                // There is nothing before turn 1 phase 1, so restart the phase
+                newPhase = 1;
+            } else {
+                newPhase = game.setupInformation.phases.length;
+                turnNumber -= 1;
+            }
+        }
+
+        const newTurnInformation = generateNewTurnInformation(
+            newPhase,
+            turnNumber,
+            game.setupInformation,
+        );
+
+        if (isLeft(newTurnInformation)) {
+            return newTurnInformation;
+        }
+
+        return MakeRight({
+            turnInformation: newTurnInformation.right,
+        });
+    },
+    "back-turn": (game) => {
+        const turnInformation = game.turnInformation;
+
+        const newPhase = 1;
+
+        const turnNumber = Math.max(turnInformation.turnNumber - 1, 1);
+
+        const newTurnInformation = generateNewTurnInformation(
+            newPhase,
+            turnNumber,
+            game.setupInformation,
+        );
+
+        if (isLeft(newTurnInformation)) {
+            return newTurnInformation;
+        }
+
+        return MakeRight({
+            turnInformation: newTurnInformation.right,
+        });
+    },
+    "forward-phase": (game) => {
+        if (isAtTurnLimit(game)) {
+            return MakeRight(game);
+        }
+
+        const turnInformation = game.turnInformation;
+
+        const { newPhase, turnNumber } =
+            turnInformation.currentPhase == game.setupInformation.phases.length
+                ? { newPhase: 1, turnNumber: turnInformation.turnNumber + 1 }
+                : {
+                      newPhase: turnInformation.currentPhase + 1,
+                      turnNumber: turnInformation.turnNumber,
+                  };
+
+        const newTurnInformation = generateNewTurnInformation(
+            newPhase,
+            turnNumber,
+            game.setupInformation,
+        );
+
+        if (isLeft(newTurnInformation)) {
+            return newTurnInformation;
+        }
+
+        return MakeRight({
+            turnInformation: newTurnInformation.right,
+        });
+    },
+    "forward-turn": (game) => {
+        const turnInformation = game.turnInformation;
+
+        const newPhase = 1;
+        const turnNumber = Math.min(
+            turnInformation.turnNumber + 1,
+            game.setupInformation.maxTurns ?? Infinity,
+        );
+
+        const newTurnInformation = generateNewTurnInformation(
+            newPhase,
+            turnNumber,
+            game.setupInformation,
+        );
+
+        if (isLeft(newTurnInformation)) {
+            return newTurnInformation;
+        }
+
+        return MakeRight({
+            turnInformation: newTurnInformation.right,
+        });
+    },
+};
 
 export function generateNewTurnInformation(
     phase: Game["turnInformation"]["currentPhase"],
